First Commit
This lives an idea for a site to keep track of personal anime watchings. Note which ones I'm watching, some sort of review / rating for it. Will be using the WebDev project as a reference. Will also have many comments to serve as reminders for what I learned and how and why I did this or that.